backend/requirements/common.txt
django==3.2
djangorestframework
Pillow
django-dotenv
mysqlclient